cloud/backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.32.*
sqlalchemy==2.0.*
alembic==1.13.*
pydantic==2.9.*
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.*
cryptography==43.*
python-multipart==0.0.*
